Create_Multiple_VMs_modules/requirements.txt
msrestazure
azure-mgmt-resource
azure-mgmt-core
azure-identity
azure-mgmt-network
azure-core